add confirmedIfNecessary method.
This method will call the callback it is provided, if none of the
conditions for nagging the user is met, or if the user confirms
the window.confirm prompt we provided.

It is a manual way of triggering the on before unload behaviour,
which we needed in a place where we did client-side logouts by
unsetting a cookie.